Delegate Req.Steps.compressed,decompress_body to Req.Decompress
